requirements.txt
astroid==2.0.4
bcrypt==3.1.4
blinker==1.4
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
docopt==0.6.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-ReCaptcha==0.4.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Migrate==2.4.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.16.3
marshmallow-sqlalchemy==0.15.0
mccabe==0.6.1
Pillow==5.4.1
pipreqs==0.4.9
pycparser==2.19
pylint==2.1.1
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.33.8
sqlalchemy-views==0.2.2
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
yarg==0.1.9
